apps/asset-worker/pyproject.toml
[project]
name = "presentation-studio-asset-worker"
version = "0.1.0"
description = "Pillow-based asset worker for presentation-studio-mcp"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
